+++ restful-test-instances/CHANGELOG.md
@@ -8,9 +8,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 
-- The tests for other organisation repository methods on the test containers.
+- The tests for other organisation service methods.
 
-## [0.3.0] - yyyy-mm-dd
+## [0.4.0] - 2020-06-11
+
+### Added
+
+The tests for other organisation repository methods on the test containers.
+
+## [0.3.0] - 2020-05-02
 
 ### Added